ssml/http_interface/requirements.txt
requests
spacy
librosa
numpy==1.21
xgboost==1.4.2
scikit-learn==0.24.2
audobject
audiofile
audinterface
audonnx
SoundFile
torch
transformers
Flask
google-cloud
google-cloud-speech
google-cloud-texttospeech
httplib2
SpeechRecognition
